pythonProject/requirements.txt
aiogram>=3.22
magic-filter>=1.0,<2
aiohttp>=3.8
python-dotenv==1.0.1
xlrd==2.0.1
openpyxl==3.1.5
pandas==2.3.3
numpy>=2.2,<3
pytz==2024.1
python-telegram-bot>=20.7,<22
psutil==5.9.8
Pillow>=10.2,<12
certifi
# для починки «битых» XLS через COM под Windows:
pywin32==306; sys_platform == "win32"

